% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{activityHPD}
\alias{activityHPD}
\title{Activity highest posterior density interval estimates (activityHPD)}
\usage{
activityHPD(model, sample = 1000, prob = 0.5, scale = "2pi")
}
\arguments{
\item{model}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{sample}{Number of posterior samples for which to calculate 95\% HDIs}

\item{prob}{Value to use for probability density mass; default=0.50}

\item{scale}{Scale of the data for plotting, either "2pi" for 0,2pi or "pi" for -pi,pi; default="2pi"}
}
\value{
Returns object of class \code{GCMMestimate} with list including estimated peak activity times, maximum activity
probability density, HPD interval, HPD duration, and number of activity peaks
}
\description{
Calculate activity highest posterior density interval (activityHPD), HPD duration, number of activity peaks,
peak activity times and maximum activity probability density for a given probability density mass
}
