% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APDfunctions.R
\name{APDpointplot}
\alias{APDpointplot}
\title{Activity Probability Density Point Plot}
\usage{
APDpointplot(
  model,
  time,
  ymax = "NULL",
  scale = "2pi",
  cex = 2.5,
  col = "lightseagreen",
  axisunits = "radians"
)
}
\arguments{
\item{model}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{time}{Time point for which to plot predicted activity probability density}

\item{ymax}{Upper limit of y axis}

\item{scale}{Scale for which to plot the activity curve, either "2pi" for 0,2pi or "pi" for -pi, pi; default="2pi"}

\item{cex}{Size of plotted point; default=2.5}

\item{col}{Colour of plotted point}

\item{axisunits}{Units for xaxis}
}
\value{
No return value; prints plot of activity curve with activity probability density prediction at the specified time point and returns dataframe of time points and activity probability density
}
\description{
Plot of GCMM activity with predicted activity probability density at particular time points
}
