% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{calcAPD}
\alias{calcAPD}
\title{Calculate activity probability density}
\usage{
calcAPD(x, curve)
}
\arguments{
\item{x}{Value in radians for which to predict probability density}

\item{curve}{Temporal data to predict density from}
}
\value{
Returns activity probability density value
}
\description{
Support function for calculating activity probability density at a specified time from an activity curve
}
