% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{calcprop}
\alias{calcprop}
\title{Calculate proportions of circular variable within an interval}
\usage{
calcprop(x, p1, p2)
}
\arguments{
\item{x}{Vector of data}

\item{p1}{Number identifying start of interval}

\item{p2}{Number identifying end of interval}
}
\value{
Returns proportion of the vector within the interval
}
\description{
Support function that calculates proportion of a vector of circular data within an interval
}
