% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{circaxis}
\alias{circaxis}
\title{Axis labels for circular temporal data plots}
\usage{
circaxis(axisunits = c("radians", "time", "sun", "none"))
}
\arguments{
\item{axisunits}{Scale to use for the xaxis, either "radians", "time", "sun", or "none"; default="radians"}
}
\value{
Prints axis
}
\description{
Support function for axis labels for circular plots
}
