% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{combineMCMC}
\alias{combineMCMC}
\title{Combine MCMC chains for posterior simulations}
\usage{
combineMCMC(model)
}
\arguments{
\item{model}{Object of class GCMM containing output from GCMM function}
}
\value{
Returns a list of MCMC chains
}
\description{
Support function that extracts MCMC chains for creating posterior simulations of activity curves
}
