% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{convertRad}
\alias{convertRad}
\title{Convert Radians Scale}
\usage{
convertRad(x, to)
}
\arguments{
\item{x}{Vector of observations in radians}

\item{to}{Character vector to specify conversion, either "2pi" for (0, 2pi), "pi" for (-pi, pi), or "hours" for hours}
}
\value{
Returns vector of observations on desired scale
}
\description{
Converts scale of observations in radians to (0,2pi) or (-pi,pi) or hours
}
\examples{
Rad2<-convertRad(redfoxsample$Radians,to="2pi")
}
