% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCMM.R
\name{exampleGCMM}
\alias{exampleGCMM}
\title{Executable example of GCMM function}
\usage{
exampleGCMM()
}
\value{
Prints message with example of GCMM function using data included in the package
}
\description{
Example of applying generalized circular mixed effect mixture model with activityGCMM
using data included in the package
}
\examples{
{ exampleGCMM() }
}
