% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{extractparam}
\alias{extractparam}
\title{Extract parameters for posterior simulations}
\usage{
extractparam(model, x)
}
\arguments{
\item{model}{Object of class GCMM containing output from GCMM function}

\item{x}{Name of parameter to be extracted}
}
\value{
Returns posterior samples of the parameter
}
\description{
Support function that extracts parameter estimates for creating posterior simulations of activity curves
}
