% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{GCMMpeaksplot}
\alias{GCMMpeaksplot}
\title{Plot estimated time of activity peaks}
\usage{
GCMMpeaksplot(x, scale = "2pi", ymax = "NULL")
}
\arguments{
\item{x}{Object of class \code{GCMMestimate} with output from \code{\link{activityHPD}} function}

\item{scale}{Scale for plotting the activity curve, either "2pi" for 0,2pi or "pi" for -pi,pi; default="2pi"}

\item{ymax}{Value for upper limit of y axis}
}
\value{
No return value; prints plot activity curve and peak activity time
}
\description{
Plot mean GCMM activity curve with peak activity time from \code{\link{activityHPD}} function
}
