% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{GCMMprob}
\alias{GCMMprob}
\title{Probability of Activity during Time Period}
\usage{
GCMMprob(model, timestart, timeend, sample = 1000)
}
\arguments{
\item{model}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{timestart}{Start of time period (in radians)}

\item{timeend}{End of time period (in radians)}

\item{sample}{Number of posterior samples from which to calculate the HDI; default=1000}
}
\value{
Returns object of class \code{GCMMestimate} containing list with summary of output with mean and 95\% HDI and the posterior distribution of predicted activity probability
}
\description{
Calculate activity probability estimates from a GCMM model during a specific period of time
}
