% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{GCMMsims}
\alias{GCMMsims}
\title{Create GCMM simulations}
\usage{
GCMMsims(PD, s)
}
\arguments{
\item{PD}{Posterior samples of GCMM parameters; output from GCMMsimsparams function}

\item{s}{Index value for running simulations}
}
\value{
Returns a vector of data simulated from the GCMM mixture
}
\description{
Support function that creates posterior simulations of GCMM activity curves
}
