% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{GCMMsimsparams}
\alias{GCMMsimsparams}
\title{Extract GCMM parameters for running GCMM simulations}
\usage{
GCMMsimsparams(model, sample)
}
\arguments{
\item{model}{Object of class GCMM containing output from GCMM function}

\item{sample}{Number of posterior samples}
}
\value{
Returns a vector posterior draws
}
\description{
Support function that extracts posterior samples of GCMM parameters for posterior simulations of GCMM activity curves
}
