% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{HDI}
\alias{HDI}
\title{Calculate highest density interval}
\usage{
HDI(x, prob = 0.95)
}
\arguments{
\item{x}{Vector of data}

\item{prob}{Value for probability mass of HDI; default=95\%}
}
\value{
Returns matrix of the mean and upper and lower bounds of the HDI
}
\description{
Calculates the highest density interval
}
