% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{HPDoverlap}
\alias{HPDoverlap}
\title{Activity HPD Overlap}
\usage{
HPDoverlap(model1, model2, prob = 0.5, sample = 1000, scale = "2pi")
}
\arguments{
\item{model1}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{model2}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{prob}{Value of probability density mass for activityHPD; default=0.50}

\item{sample}{Number of posterior samples for calculating 95\% HDI; default=1000}

\item{scale}{Scale for plotting the activity curve, either "2pi" for 0,2pi or "pi" for -pi,pi; default="2pi"}
}
\value{
Returns object of class  \code{GCMMestimate} containing list of posterior distributions and summary information of mean and 95\% HDI
}
\description{
Calculate whether there is overlap between two HPDs, the amount of overlap, and the probability of activity
during the HPD of the other
}
