% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanssample}
\alias{humanssample}
\title{Sample data of camera trap observations of humans}
\format{
Dataframes with 3 variables
Radians Time of observations, in radians (0 to 2pi)
CameraTrapID Variable identifying camera traps
SamplingPeriod Variable identifying sampling period during which camera traps were recording
}
\source{
\ Campbell L.A.D. 2017
}
\usage{
humanssample
}
\description{
Example dataset for fitting circular mixed effect mixture models with activityGCMM package
}
\details{
Sample data of camera trap observations of humans
}
\examples{
data(humanssample)
\dontrun{ GCMM(data=humanssample$Radians, RE1=humanssample$SamplingPeriod, 
    scale=c("2pi"), family="vonmises", autojags=TRUE, thin=3) }

}
\keyword{datasets}
