% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{PDS}
\alias{PDS}
\title{Posterior distribution summaries and support}
\usage{
PDS(x)
}
\arguments{
\item{x}{Vector of data}
}
\value{
Returns matrix with data summary
}
\description{
Calculates the proportion > 0, < 0, and posterior distribution support (PDS)
}
