% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{peaksPDplot}
\alias{peaksPDplot}
\title{Plot estimated number of activity peaks}
\usage{
peaksPDplot(x, col = "cyan4")
}
\arguments{
\item{x}{Object of class \code{GCMMestimate} with output from \code{\link{activityHPD}} function}

\item{col}{Colour for plot}
}
\value{
No return value; prints histogram plot of posterior estimates
}
\description{
Plot of posterior samples of estimated number of activity peaks from \code{\link{activityHPD}} function
}
