% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityProbabilityFunctions.R
\name{plotactivityHPD}
\alias{plotactivityHPD}
\title{Plot activity curve with activityHPDs}
\usage{
plotactivityHPD(
  model,
  prob = c(0.75, 0.5, 0.25),
  col = c("lightseagreen", "aquamarine3", "aquamarine"),
  scale = "2pi",
  ymax = "NULL",
  axisunits = "radians"
)
}
\arguments{
\item{model}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{prob}{Vector of values of probability density mass for activityHPD; default=c(0.50,0.25)}

\item{col}{Vector of colours for plot}

\item{scale}{Scale for plotting the activity curve, either "2pi" for 0,2pi or "pi" for -pi,pi; default="2pi"}

\item{ymax}{Value for upper limit of y axis}

\item{axisunits}{Units for x axis, either "radians", "time", "solar", "sun", or "none"; default="radians"}
}
\value{
No return value; prints plot of activity curve with activityHPDs
}
\description{
Plot GCMM activity curve with activityHPDs
}
