% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyfunctions.R
\name{posteriorhistplot}
\alias{posteriorhistplot}
\title{Plot histogram of posterior distribution}
\usage{
posteriorhistplot(model, param, col = "cyan4")
}
\arguments{
\item{model}{Object of class \code{GCMM} with output from \code{\link{GCMM}} function}

\item{param}{Parameter for which to plot the posterior samples}

\item{col}{Colour of histogram}
}
\value{
Returns matrix with the mean and HDI of activity probability density estimated from both GCMM models at the peak activity time of the other. Posterior distributions of activity probability density and peak activity times for both GCMM models are also saved.
}
\description{
Plot histogram of posterior samples
}
