% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{progress}
\alias{progress}
\title{Report function progress}
\usage{
progress(s, sample)
}
\arguments{
\item{s}{Number of iteration in the loop}

\item{sample}{Total number of iterations in the loop}
}
\value{
No return value; prints progress of function at 10\% intervals
}
\description{
Support function that prints progress of functions with long computation times at 10\% intervals
}
