% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redfoxsample}
\alias{redfoxsample}
\title{Sample data of camera trap observations of red fox}
\format{
Dataframes with 3 variables
Radians Time of observations, in radians (0 to 2pi)
CameraTrapID Variable identifying camera traps
SamplingPeriod Variable identifying sampling period during which camera traps were recording
}
\source{
\ Campbell L.A.D. 2017
}
\usage{
redfoxsample
}
\description{
Example dataset for fitting circular mixed effect mixture models with AnimalAPD package
}
\details{
Sample data of camera trap observations of humans
}
\examples{
data(redfoxsample)
\dontrun{ GCMM(data=redfoxsample$Radians, RE1=redfoxsample$SamplingPeriod, 
    scale=c("2pi"), family="vonmises", autojags=FALSE,
    adapt=0, sample=300, burnin=300, thin=1, n.chains=2  ) }

}
\keyword{datasets}
