% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{samplerows}
\alias{samplerows}
\title{Sample rows of dataframe}
\usage{
samplerows(df, n)
}
\arguments{
\item{df}{Dataframe}

\item{n}{Number of samples}
}
\value{
Returns sample of dataframe with the number of specified rows
}
\description{
Support function that samples rows of data from a dataframe
}
