% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Variables}
\alias{Variables}
\title{Global variables}
\value{
Summary of model output, depending on printvars & mixture plot of components; also saves to global environment the model intercepts, concentration parameters, mixture component weights, model as a runjags object ("GCMM") which can be extended with extend.jags or monitored parameters added/dropped

out Model output

Mu Component means

P Component weights

K Component concentration

GCMMcomponents Separate mixture components

GCMMmixture Mixture
}
\description{
List of variables
}
\details{
Global variables

Variables List of variables
}
