% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportfunctions.R
\name{yMax}
\alias{yMax}
\title{Calculate y-axis limit for plotting multiple activity curves}
\usage{
yMax(models, type = "mixture")
}
\arguments{
\item{models}{List of objects of class GCMM containing output from GCMM function}

\item{type}{Identifier for whether to use maximum probability density from GCMM mixture, using "mixture", or components, using "components"; default is to use the GCMM mixture density}
}
\value{
Returns a value of the maximum probability density plus buffer space to be used as the y-axis limit in activity curve plots
}
\description{
Identifies maximum probability density for multiple activity curves to select y-axis limit when plotting multiple curves
}
